moved GoogleSignInManager to profile module
