+++ app/build.gradle
@@ -149,8 +149,8 @@ android {
     }
     packagingOptions {
         resources {
-            exclude("/META-INF/{AL2.0,LGPL2.1}")
-            exclude("META-INF/DEPENDENCIES")
+            excludes += "/META-INF/{AL2.0,LGPL2.1}"
+            excludes += "META-INF/DEPENDENCIES"
         }
     }
     testOptions {
@@ -196,10 +196,6 @@ dependencies {
     implementation libs.composeMaterial3
     implementation libs.material3WindowSizeClass
     implementation libs.splash
-    implementation libs.playServicesAuth
-    implementation libs.googleApiClient
-    implementation libs.googleApiServicesDrive
-    implementation libs.documentFile
 
     implementation libs.immutableCollections
