+++ app/src/main/java/com/babylon/wallet/android/presentation/onboarding/OnboardingViewModel.kt
@@ -8,10 +8,10 @@ import com.babylon.wallet.android.presentation.common.OneOffEventHandler
 import com.babylon.wallet.android.presentation.common.OneOffEventHandlerImpl
 import com.babylon.wallet.android.presentation.common.StateViewModel
 import com.babylon.wallet.android.presentation.common.UiState
-import com.babylon.wallet.android.presentation.googlesignin.GoogleSignInManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import rdx.works.profile.cloudbackup.GoogleSignInManager
 import timber.log.Timber
 import javax.inject.Inject
