+++ app/src/main/java/com/babylon/wallet/android/presentation/onboarding/restore/backup/RestoreFromBackupViewModel.kt
@@ -10,13 +10,13 @@ import com.babylon.wallet.android.presentation.common.OneOffEventHandlerImpl
 import com.babylon.wallet.android.presentation.common.StateViewModel
 import com.babylon.wallet.android.presentation.common.UiMessage
 import com.babylon.wallet.android.presentation.common.UiState
-import com.babylon.wallet.android.presentation.googlesignin.GoogleSignInManager
 import com.babylon.wallet.android.utils.Constants
 import com.radixdlt.sargon.Profile
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
+import rdx.works.profile.cloudbackup.GoogleSignInManager
 import rdx.works.core.sargon.isCompatible
 import rdx.works.profile.domain.ProfileException
 import rdx.works.profile.domain.backup.BackupType
