+++ profile/src/main/java/rdx/works/profile/cloudbackup/GoogleAccount.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.googlesignin
+package rdx.works.profile.cloudbackup
 
 import android.net.Uri
